Refactor comps into a separate subpackage
