+++ cheta/fetch.py
@@ -28,7 +28,7 @@ from . import (
     file_defs,
     remote_access,
 )
-from .derived.comps import ComputedMsid
+from .comps import ComputedMsid
 from .lazy import LazyDict
 from .remote_access import ENG_ARCHIVE
 from .units import Units
